The averaged top1, top5 and loss scores are incorrectly computed by method train and test.
Explanations:

Assume a dataset size of N = 100, and a batch size of bs = 30.
This makes 4 batches of size 30, 30, 30, 10.
The top1Sum and top5sum are currently computed as follows:
( s1/30 + s2/30 + s3/30 + s4/10 )/4 * 100
where si are the top1 (or top5) numbers of incorrectly classified samples.

This is incorrect because the last batch has a size smaller than 30. The
correct way to compute the score is instead:
(s1 + s2 + s3 + s4)/N * 100

This error comes from assuming that M mod bs = 0, which is not always the case.